+++ common.js
@@ -1,5 +1,0 @@
-import './common/agreement'
-import './common/legal'
-
-export { LegalAgreementCollection } from './common/agreement'
-export { LegalCollection } from './common/legal'