+++ common.ts
@@ -1,0 +1,5 @@
+import './common/agreement'
+import './common/legal'
+
+export { LegalAgreementCollection, LegalAgreement } from './common/agreement'
+export { LegalCollection, LegalDocument } from './common/legal'