+++ server.js
@@ -1,4 +1,0 @@
-import './server/agreement-server'
-import './server/legal-server'
-
-export { LegalAgreementCollection, LegalCollection } from './common'